+++ lib/src/indicators/line_scale_pulse_out.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:loading_indicator/src/decorate/decorate.dart';
 import 'package:loading_indicator/src/indicators/base/indicator_controller.dart';
 import 'package:loading_indicator/src/shape/indicator_painter.dart';
 
@@ -42,6 +43,8 @@ class _LineScalePulseOutState extends State<LineScalePulseOut>
 
   @override
   Widget build(BuildContext context) {
+    final lineWidth =
+        DecorateContext.of(context)!.decorateData.configuredStrokeWidth;
     final widgets = List<Widget>.filled(9, Container());
     for (int i = 0; i < widgets.length; i++) {
       if (i.isEven) {
@@ -53,6 +56,7 @@ class _LineScalePulseOutState extends State<LineScalePulseOut>
                 heightFactor: _animations[i ~/ 2].value,
                 child: IndicatorShapeWidget(
                   shape: Shape.line,
+                  lineWidth: lineWidth,
                   index: i ~/ 2,
                 ),
               );